Transform Docker build output to hide internals of Docker interception
This isn't for security or similar in any sense (the full details are
available through 'docker history' anyway) but there's a lot of extra
noise here that's unhelpful & confusing. It's much clearer to trim it
down to a single "Enabling HTTP Toolkit interception" line instead.

This handles that, but the veil is a little imperfect - e.g. the Step
counter numbers jump suddenly.